% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinadevfin3-package.R
\docType{package}
\name{chinadevfin3-package}
\alias{chinadevfin3}
\alias{chinadevfin3-package}
\title{chinadevfin3: An R Package for AidData's Global Chinese Development Finance 3.0 Dataset}
\description{
`chinadevfin3` is a package for working with AidData's Global Chinese Development Finance 3.0 Dataset. The dataset includes 20,985 Chinese development finance projects across 165 countries worth $1.34 trillion from 2000-2021. `chinadevfin3` provides an easy way for R users to load the dataset, and helper functions to ease common analysis workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/t-emery/chinadevfin3}
  \item \url{https://t-emery.github.io/chinadevfin3/}
  \item Report bugs at \url{https://github.com/t-emery/chinadevfin3/issues}
}

}
\author{
\strong{Maintainer}: Teal Emery \email{lte@tealinsights.com}

}
\keyword{internal}
