% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcdf3_dataset}
\alias{gcdf3_dataset}
\title{AidData Global Chinese Development Finance Dataset, Version 3.0}
\format{
A tibble with 20,985 rows and 126 variables:
\describe{
\item{Dataset}{\code{gcdf2_data_dictionary} contains definitions of all 126 variables.}

}
}
\source{
\url{https://www.aiddata.org/data/aiddatas-global-chinese-development-finance-dataset-version-3-0}
}
\usage{
gcdf3_dataset
}
\description{
The dataset captures 20,985 projects across 165 low- and middle-income countries supported by loans and grants from official sector institutions in China worth $1.34 trillion. It tracks projects over 22 commitment years (2000-2021) and provides details on the timing of project implementation over a 24-year period (2000-2023).
}
\keyword{datasets}
